Thanks John
Luke – can we make this update?  I’d suggest we just change the field so that someone can pick a name but it does nothing other than show that person.

Our next deployment to production for updates is on the 17th so we’ll see if it can get into that cycle, otherwise it may be the 1st of July.

Nicole – note we’ll have to change all the training materials to reflect this – so we need to consider if we want to delay this until we have done the entire review and catalogued all the desired changes and do one sweep of updates rather than being in a continuous update cycle.

Jodi

Jodi,

We discussed this with Mike Kolloway today and he does not want this functionality included for any of the geographies.  He is OK with allowing someone to suggest a reviewer, but not assign/choose.

What is the process to get this changed?

Thanks………….John